Refactor price service
