+++ backend/main.go
@@ -1,7 +1,8 @@
 package main
 
 import (
-	"github.com/dath-241/coin-price-be-go/routes"
+	// "github.com/dath-241/coin-price-be-go/services/price-service/routes"
+	"github.com/dath-241/coin-price-be-go/services/price-service/src/routes"
 	"github.com/gin-gonic/gin"
 )
 
@@ -15,6 +16,7 @@ func main() {
 	server := gin.Default()
 
 	routes.RegisterRoutes(server)
+	// routes.RegisterRoutes(server)
 
 	server.Run(":8080")
 	//Set up router, routs, server, websocket
