--- backend/services/price-service/models/price.go
@@ -1,4 +1,0 @@
-package models
-
-// Chua cac struct/interface de lay du lieu su dung cho services/reposistories
-type StatusCode int
