--- backend/services/price-service/reposistories/price.go
@@ -1,3 +1,0 @@
-package reposistories
-
-//Cac ham crud/query data tu db (mongodb)
