--- backend/services/price-service/reposistories/price_test.go
@@ -1,3 +1,0 @@
-package reposistories
-
-//Cac ham test connect db, test crud/query (thuc hien tren collection test cua mongodb, khong lam tren collection chinh)
