--- backend/services/price-service/services/price.go
@@ -1,3 +1,0 @@
-package services
-
-//Cac ham API (request -> response), call func cua reposistories de tuong tac voi db
