--- backend/services/price-service/services/websocket/websocket.go
@@ -1,26 +1,0 @@
-package websocket
-
-import (
-	"fmt"
-	"net/http"
-
-	"github.com/gorilla/websocket"
-)
-
-// Set up websocket
-var upgrader = websocket.Upgrader{
-	ReadBufferSize:  1024,
-	WriteBufferSize: 1024,
-}
-
-func Upgrade(w http.ResponseWriter, r *http.Request) (*websocket.Conn, error) {
-	upgrader.CheckOrigin = func(r *http.Request) bool { return true }
-
-	ws, err := upgrader.Upgrade(w, r, nil)
-	if err != nil {
-		fmt.Println(err)
-		return ws, err
-	}
-
-	return ws, nil
-}
