+++ backend/services/price-service/src/models/price.go
@@ -1,0 +1,4 @@
+package models
+
+// Chua cac struct/interface de lay du lieu su dung cho services/reposistories
+type StatusCode int
