+++ backend/services/price-service/src/reposistories/price.go
@@ -1,0 +1,3 @@
+package reposistories
+
+//Cac ham crud/query data tu db (mongodb)
