+++ backend/services/price-service/src/reposistories/price_test.go
@@ -1,0 +1,3 @@
+package reposistories
+
+//Cac ham test connect db, test crud/query (thuc hien tren collection test cua mongodb, khong lam tren collection chinh)
