+++ backend/services/price-service/src/services/price.go
@@ -1,0 +1,3 @@
+package services
+
+//Cac ham API (request -> response), call func cua reposistories de tuong tac voi db
